Use tabulator key to move focus into command toolbar
Instead of (partially) moving the focus automatically into the command
toolbar (e.g. when adding a component to the schematic), the focus
always stays in the graphics view. But the tabulator key now moves
the focus into the command toolbar in any tool.

Reason: Having the focus in the toolbar sometimes breaks keyboard
shortcuts (e.g. 'R' for rotate) because some toolbar widgets contain
text input fields which consume these shortcuts - therefore this
shouldn't happen automatically. Downside is that in some tools the
user now has to press 'tab' manually to start editing for example the
value of a text label, but at least this concept is now consistent
across all tools.